biglake-iceberg-pipeline/test_data/thelook_ecommerce/requirements.txt
faker>=33.0.0
google-cloud-bigquery>=3.0.0
